feat: setup failing tests
